+++ tests/manual/test.py
@@ -1,16 +1,18 @@
 from pyespn import PYESPN
 player_id = 278 # Jimmy Smith, Goat
-season = 2025
+season = 2024
 pick = 8
 pick_round = 2
 
 
 if __name__ == '__main__':
 
-    espn = PYESPN('mlb')
+    espn = PYESPN('nfl')
     espn.load_season_rosters(season=season)
-    espn.load_season_schedule(season=season, load_only_current_week=True)  #, load_game_odds=True)
-
+    espn.load_season_schedule(season=season)  #, load_game_odds=True)
+    for week in espn.league.schedules[season].weeks:
+        week.load_event_officials()
+        week.load_event_broadcasts()
     espn.league.load_regular_season_schedule(season=season, load_game_odds=True)
     #espn.load_season_rosters(season=season)
     #espn.teams[0].load_season_roster_box_score(season=season)